Remove code duplication, enhance tests
